refactor: 여러개의 룰을 가지는 OmokRule 룰 구현체 생성
