+++ app/src/main/java/woowacourse/omok/domain/rule/OmokRule.kt
@@ -1,6 +1,6 @@
+
 package woowacourse.omok.domain.rule
 
-import rule.BlackRenjuRule
 import woowacourse.omok.domain.omokboard.OmokBoard
 import woowacourse.omok.domain.placeresult.PlaceResult
 import woowacourse.omok.domain.player.PlayerStone
@@ -10,23 +10,4 @@ interface OmokRule {
         omokBoard: OmokBoard,
         playerStone: PlayerStone,
     ): PlaceResult
-
-    companion object {
-        private val board = OmokBoard()
-
-        val whiteRules: List<OmokRule> get() =
-            listOf(
-                InvalidPositionRule(),
-                DrawRule(),
-                WinningRule(),
-            )
-
-        val blackRules: List<OmokRule> get() =
-            listOf(
-                InvalidPositionRule(),
-                ExternalRenjuRule(BlackRenjuRule(board.width, board.height)),
-                DrawRule(),
-                WinningRule(),
-            )
-    }
 }
